Plutus.Script.Utils.Ada

Description

Functions for working with Ada in Template Haskell.

Documentation

newtype Ada

ADA, the special currency on the Cardano blockchain. The unit of Ada is Lovelace, and 1M Lovelace is one Ada. See note [Currencies] in TH.

Constructors

Lovelace

Fields

getLovelace :: Integer

getAda :: Ada -> Micro

Get the amount of Ada (the unit of the currency Ada) in this Ada value.

adaSymbol :: CurrencySymbol

The CurrencySymbol of the Ada currency.

adaToken :: TokenName

The TokenName of the Ada currency.

Constructors

fromValue :: Value -> Ada

Get the Ada in the given Value.

toValue :: Ada -> Value

Create a Value containing only the given Ada.

lovelaceOf :: Integer -> Ada

Create Ada representing the given quantity of Lovelace (the unit of the currency Ada).

adaOf :: Micro -> Ada

Create Ada representing the given quantity of Ada (1M Lovelace).

lovelaceValueOf :: Integer -> Value

A Value with the given amount of Lovelace (the currency unit).

lovelaceValueOf == toValue . lovelaceOf

adaValueOf :: Micro -> Value

A Value with the given amount of Ada (the currency unit).

adaValueOf == toValue . adaOf

Num operations

divide :: Ada -> Ada -> Ada

Divide one Ada value by another.

Etc.

isZero :: Ada -> Bool

Check whether an Ada value is zero.