Plutus.Script.Utils.Scripts

Description

This module contains functions related to versioning scripts and BuiltinData, or more specifially, Datums and Redeemers. These functions do not depend on a particular version of Plutus.

Plutus language versioning

data Language

Constructors

PlutusV1

PlutusV2

data Versioned script

A script of some kind with its Plutus language version

Constructors

Versioned

Fields

unversioned :: script

version :: Language

Script hashing

scriptHash :: Versioned Script -> ScriptHash

Hash a Versioned Script

validatorHash :: Versioned Validator -> ValidatorHash

Hash a Versioned Validator script.

mintingPolicyHash :: Versioned MintingPolicy -> MintingPolicyHash

Hash a Versioned MintingPolicy script.

stakeValidatorHash :: Versioned StakeValidator -> StakeValidatorHash

Hash a Versioned StakeValidator script.

Script utilities

scriptCurrencySymbol :: Versioned MintingPolicy -> CurrencySymbol

The CurrencySymbol of a MintingPolicy.

Script data hashes

data Datum

data DatumHash

data Redeemer

data RedeemerHash

datumHash :: Datum -> DatumHash

Hash a 'PV1.Datum builtin data.

redeemerHash :: Redeemer -> RedeemerHash

Hash a Redeemer builtin data.

dataHash :: BuiltinData -> BuiltinByteString

Hash a BuiltinData

Orphan instances

Serialise Language

Instance details

Methods

encode :: Language -> Encoding

decode :: Decoder s Language

encodeList :: [Language] -> Encoding

decodeList :: Decoder s [Language]

Pretty Language

Instance details

Methods

pretty :: Language -> Doc ann

prettyList :: [Language] -> Doc ann