Plutus.Script.Utils.Typed

Documentation

type UntypedValidator = BuiltinData -> BuiltinData -> BuiltinData -> ()

type UntypedMintingPolicy = BuiltinData -> BuiltinData -> ()

type UntypedStakeValidator = BuiltinData -> BuiltinData -> ()

class ValidatorTypes (a :: Type)

A class that associates a type standing for a connection type with two types, the type of the redeemer and the data script for that connection type.

Associated Types

type RedeemerType a :: Type

The type of the redeemers of this connection type.

type RedeemerType a = ()

type DatumType a :: Type

The type of the data of this connection type.

type DatumType a = ()

data TypedValidator (a :: Type)

A typed validator script with its ValidatorScript and Address.

Constructors

TypedValidator

Fields

tvValidator :: Versioned Validator

tvValidatorHash :: ValidatorHash

tvForwardingMPS :: Versioned MintingPolicy

tvForwardingMPSHash :: MintingPolicyHash

The hash of the minting policy that checks whether the validator is run in this transaction

validatorHash :: TypedValidator a -> ValidatorHash

The hash of the validator.

validatorCardanoAddress :: NetworkId -> TypedValidator a -> AddressInEra BabbageEra

The address of the validator.

validatorCardanoAddressAny :: NetworkId -> TypedValidator a -> AddressAny

validatorAddress :: TypedValidator a -> Address

The address of the validator.

validatorScript :: TypedValidator a -> Validator

The unversioned validator script itself.

vValidatorScript :: TypedValidator a -> Versioned Validator

The validator script itself.

forwardingMintingPolicy :: TypedValidator a -> MintingPolicy

The unversioned minting policy that forwards all checks to the instance's validator

vForwardingMintingPolicy :: TypedValidator a -> Versioned MintingPolicy

The minting policy that forwards all checks to the instance's validator

forwardingMintingPolicyHash :: TypedValidator a -> MintingPolicyHash

Hash of the minting policy that forwards all checks to the instance's validator

generalise :: forall a. TypedValidator a -> TypedValidator Any

Generalise the typed validator to one that works with the Data type.

data Any

data Language

Constructors

PlutusV1

PlutusV2

data Versioned script

A script of some kind with its Plutus language version

Constructors

Versioned

Fields

unversioned :: script

version :: Language

class UnsafeFromData sc => IsScriptContext sc where

Minimal complete definition

Nothing

Methods

mkUntypedValidator :: (UnsafeFromData d, UnsafeFromData r) => (d -> r -> sc -> Bool) -> UntypedValidator

Converts a custom datum and redeemer from a validator function to an untyped validator function. See Note [Scripts returning Bool].

Here's an example of how this function can be used:

  import PlutusTx qualified
  import Plutus.V2.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V2.Scripts (mkUntypedValidator)

  newtype MyCustomDatum = MyCustomDatum Integer
  PlutusTx.unstableMakeIsData ''MyCustomDatum
  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkValidator :: MyCustomDatum -> MyCustomRedeemer -> Plutus.ScriptContext -> Bool
  mkValidator _ _ _ = True

  validator :: Plutus.Validator
  validator = Plutus.mkValidatorScript
      $$(PlutusTx.compile [|| wrap ||])
   where
      wrap = mkUntypedValidator mkValidator

Here's an example using a parameterized validator:

  import PlutusTx qualified
  import Plutus.V2.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V2.Scripts (mkUntypedValidator)

  newtype MyCustomDatum = MyCustomDatum Integer
  PlutusTx.unstableMakeIsData ''MyCustomDatum
  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkValidator :: Int -> MyCustomDatum -> MyCustomRedeemer -> Plutus.ScriptContext -> Bool
  mkValidator _ _ _ _ = True

  validator :: Int -> Plutus.Validator
  validator i = Plutus.mkValidatorScript
      $$(PlutusTx.compile [|| wrap . mkValidator ||]) applyCode PlutusTx.liftCode i
   where
      wrap = mkUntypedValidator

For debugging purpose, it may be of interest to know that in the default implementation, the parameters are decoded in the order they appear (data, redeemer and then script context). A log trace is generated after each successfully decoded parameter. Thus, if a parameter can't be decoded, the culprit is the first parameter in the list that doesn't appear as successfully decoded in the log trace.

mkUntypedStakeValidator :: UnsafeFromData r => (r -> sc -> Bool) -> UntypedStakeValidator

Converts a custom redeemer from a stake validator function to an untyped stake validator function. See Note [Scripts returning Bool].

Here's an example of how this function can be used:

  import PlutusTx qualified
  import Plutus.V1.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V1.Scripts (mkUntypedStakeValidator)

  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkStakeValidator :: MyCustomRedeemer -> ScriptContext -> Bool
  mkStakeValidator _ _ = True

  validator :: Plutus.Validator
  validator = Plutus.mkStakeValidatorScript
      $$(PlutusTx.compile [|| wrap ||])
   where
      wrap = mkUntypedStakeValidator mkStakeValidator

For debugging purpose, it may be of interest to know that in the default implementation, the parameters are decoded in the order they appear (redeemer and then script context). A log trace is generated after each successfully decoded parameter. Thus, if a parameter can't be decoded, the culprit is the first parameter in the list that doesn't appear as successfully decoded in the log trace.

mkUntypedMintingPolicy :: UnsafeFromData r => (r -> sc -> Bool) -> UntypedMintingPolicy

Converts a custom redeemer from a minting policy function to an untyped minting policy function. See Note [Scripts returning Bool].

Here's an example of how this function can be used:

  import PlutusTx qualified
  import Plutus.V1.Ledger.Scripts qualified as Plutus
  import Plutus.Script.Utils.V1.Scripts (mkUntypedMintingPolicy)

  newtype MyCustomRedeemer = MyCustomRedeemer Integer
  PlutusTx.unstableMakeIsData ''MyCustomRedeemer

  mkMintingPolicy :: MyCustomRedeemer -> ScriptContext -> Bool
  mkMintingPolicy _ _ = True

  validator :: Plutus.Validator
  validator = Plutus.mkMintingPolicyScript
      $$(PlutusTx.compile [|| wrap ||])
   where
      wrap = mkUntypedMintingPolicy mkMintingPolicy

For debugging purpose, it may be of interest to know that in the default implementation, the parameters are decoded in the order they appear (redeemer and then script context). A log trace is generated after each successfully decoded parameter. Thus, if a parameter can't be decoded, the culprit is the first parameter in the list that doesn't appear as successfully decoded in the log trace.

type ScriptContextV1 = ScriptContext

type ScriptContextV2 = ScriptContext