Plutus.Script.Utils.V1.Contexts

Documentation

data ScriptContext

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

scriptContextPurpose :: ScriptPurpose

data ScriptPurpose

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding StakingCredential

Certifying DCert

data TxInInfo

Constructors

TxInInfo

Fields

txInInfoOutRef :: TxOutRef

txInInfoResolved :: TxOut

data TxInfo

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

txInfoOutputs :: [TxOut]

txInfoFee :: Value

txInfoMint :: Value

txInfoDCert :: [DCert]

txInfoWdrl :: [(StakingCredential, Integer)]

txInfoValidRange :: POSIXTimeRange

txInfoSignatories :: [PubKeyHash]

txInfoData :: [(DatumHash, Datum)]

txInfoId :: TxId

newtype TxId

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data TxOut

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatumHash :: Maybe DatumHash

data TxOutRef

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

findContinuingOutputs :: ScriptContext -> [Integer]

findDatum :: DatumHash -> TxInfo -> Maybe Datum

findDatumHash :: Datum -> TxInfo -> Maybe DatumHash

findOwnInput :: ScriptContext -> Maybe TxInInfo

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

fromSymbol :: CurrencySymbol -> ValidatorHash

getContinuingOutputs :: ScriptContext -> [TxOut]

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

ownHash :: ScriptContext -> ValidatorHash

ownHashes :: ScriptContext -> (ValidatorHash, DatumHash)

pubKeyOutput :: TxOut -> Maybe PubKeyHash

pubKeyOutputsAt :: PubKeyHash -> TxInfo -> [Value]

scriptOutputsAt :: ValidatorHash -> TxInfo -> [(DatumHash, Value)]

spendsOutput :: TxInfo -> TxId -> Integer -> Bool

txSignedBy :: TxInfo -> PubKeyHash -> Bool

valueLockedBy :: TxInfo -> ValidatorHash -> Value

valueProduced :: TxInfo -> Value

valueSpent :: TxInfo -> Value

outputsAt :: Address -> TxInfo -> [Value]

Get the values paid to a public key address by a pending transaction.

valuePaidTo :: TxInfo -> Address -> Value

Get the total value paid to a public key address by a pending transaction.