Plutus.Script.Utils.V1.Scripts

Description

This module contains functions related to the computation of script hashes for PlutusV1.

Script hashes

data Validator

Validator is a wrapper around Scripts which are used as validators in transaction outputs.

data ValidatorHash

Script runtime representation of a Digest SHA256.

data MintingPolicy

MintingPolicy is a wrapper around Scripts which are used as validators for minting constraints.

data MintingPolicyHash

Script runtime representation of a Digest SHA256.

data StakeValidator

StakeValidator is a wrapper around Scripts which are used as validators for withdrawals and stake address certificates.

data StakeValidatorHash

Script runtime representation of a Digest SHA256.

fromCardanoHash :: ScriptHash -> ScriptHash

Transform a Cardano Script hash in a Plutus Script hash

validatorHash :: Validator -> ValidatorHash

Hash a Validator script.

mintingPolicyHash :: MintingPolicy -> MintingPolicyHash

Hash a MintingPolicy script.

stakeValidatorHash :: StakeValidator -> StakeValidatorHash

Hash a StakeValidator script.

scriptHash :: Script -> ScriptHash

Hash a Script

Script utilities

scriptCurrencySymbol :: MintingPolicy -> CurrencySymbol

The CurrencySymbol of a MintingPolicy.

toCardanoApiScript :: Script -> Script PlutusScriptV1

Convert a Script to a 'cardano-api' script.

For why we depend on `cardano-api`, see note [Hash computation of datums, redeemers and scripts]