Plutus.Script.Utils.V1.Tx

Documentation

scriptTxOut :: Validator -> Value -> Datum -> TxOut

Create a transaction output locked by a validator script and attach the given data script.

scriptAddressTxOut :: Address -> Value -> Datum -> TxOut

Create a transaction output locked by a validator script hash with the given data script attached.