Plutus.Script.Utils.V1.Typed.Scripts

Documentation

mkForwardingMintingPolicy :: ValidatorHash -> MintingPolicy

A minting policy that checks whether the validator script was run in the minting transaction.

mkForwardingStakeValidator :: ValidatorHash -> StakeValidator

A stake validator that checks whether the validator script was run in the right transaction.

module Plutus.Script.Utils.V1.Typed.Scripts.Validators

data Validator

data MintingPolicy

data StakeValidator

data TypedScriptTxOut a

A TxOut tagged by a phantom type: and the connection type of the output.

data TypedScriptTxOutRef a

A TxOutRef tagged by a phantom type: and the connection type of the output.

makeTypedScriptTxOut :: forall out. (ToData (DatumType out), FromData (DatumType out)) => TypedValidator out -> DatumType out -> Value -> TypedScriptTxOut out

Create a TypedScriptTxOut from a correctly-typed data script, an address, and a value.

typeScriptTxOut :: forall out m. (FromData (DatumType out), ToData (DatumType out), MonadError ConnectionError m) => TypedValidator out -> TxOutRef -> TxOut -> Datum -> m (TypedScriptTxOut out)

Create a TypedScriptTxOut from an existing TxOut by checking the types of its parts.

typeScriptTxOutRef :: forall out m. (FromData (DatumType out), ToData (DatumType out), MonadError ConnectionError m) => TypedValidator out -> TxOutRef -> TxOut -> Datum -> m (TypedScriptTxOutRef out)

Create a TypedScriptTxOut from an existing TxOut by checking the types of its parts.