Plutus.Script.Utils.V2.Typed.Scripts

Documentation

mkForwardingMintingPolicy :: ValidatorHash -> MintingPolicy

A minting policy that checks whether the validator script was run in the minting transaction.

mkForwardingStakeValidator :: ValidatorHash -> StakeValidator

A stake validator that checks whether the validator script was run in the right transaction.

module Plutus.Script.Utils.V2.Typed.Scripts.Validators

data Validator

Validator is a wrapper around Scripts which are used as validators in transaction outputs.

data MintingPolicy

MintingPolicy is a wrapper around Scripts which are used as validators for minting constraints.

data StakeValidator

StakeValidator is a wrapper around Scripts which are used as validators for withdrawals and stake address certificates.

data TypedScriptTxOut a

A TxOut tagged by a phantom type: and the connection type of the output.

Constructors

(FromData (DatumType a), ToData (DatumType a)) => TypedScriptTxOut

Fields

tyTxOutTxOut :: TxOut

tyTxOutData :: DatumType a

data TypedScriptTxOutRef a

A TxOutRef tagged by a phantom type: and the connection type of the output.

Constructors

TypedScriptTxOutRef

Fields

tyTxOutRefRef :: TxOutRef

tyTxOutRefOut :: TypedScriptTxOut a

typeScriptTxOut :: forall out m. (FromData (DatumType out), ToData (DatumType out), MonadError ConnectionError m) => TypedValidator out -> TxOutRef -> TxOut -> Datum -> m (TypedScriptTxOut out)

Create a TypedScriptTxOut from an existing TxOut by checking the types of its parts.

typeScriptTxOutRef :: forall out m. (FromData (DatumType out), ToData (DatumType out), MonadError ConnectionError m) => TypedValidator out -> TxOutRef -> TxOut -> Datum -> m (TypedScriptTxOutRef out)

Create a TypedScriptTxOut from an existing TxOut by checking the types of its parts.

data ConnectionError

An error we can get while trying to type an existing transaction part.

Constructors

WrongValidatorAddress Address Address

WrongOutType WrongOutTypeError

WrongValidatorType String

WrongRedeemerType BuiltinData

WrongDatumType BuiltinData

NoDatum TxOutRef DatumHash

UnknownRef TxOutRef