Plutus.Script.Utils.V2.Typed.Scripts.Validators

Documentation

type UntypedValidator = BuiltinData -> BuiltinData -> BuiltinData -> ()

class ValidatorTypes (a :: Type)

A class that associates a type standing for a connection type with two types, the type of the redeemer and the data script for that connection type.

Associated Types

type RedeemerType a :: Type

The type of the redeemers of this connection type.

type RedeemerType a = ()

type DatumType a :: Type

The type of the data of this connection type.

type DatumType a = ()

type ValidatorType (a :: Type) = DatumType a -> RedeemerType a -> ScriptContext -> Bool

The type of validators for the given connection type.

data TypedValidator (a :: Type)

A typed validator script with its ValidatorScript and Address.

mkTypedValidator

Arguments

:: CompiledCode (ValidatorType a)
Validator script (compiled)

-> CompiledCode (ValidatorType a -> UntypedValidator)
A wrapper for the compiled validator

-> TypedValidator a

Make a TypedValidator from the CompiledCode of a validator script and its wrapper.

mkTypedValidatorParam

Arguments

:: forall a param. Lift DefaultUni param

=> CompiledCode (param -> ValidatorType a)
Validator script (compiled)

-> CompiledCode (ValidatorType a -> UntypedValidator)
A wrapper for the compiled validator

-> param
The extra paramater for the validator script

-> TypedValidator a

Make a TypedValidator from the CompiledCode of a parameterized validator script and its wrapper.

validatorHash :: TypedValidator a -> ValidatorHash

The hash of the validator.

validatorAddress :: TypedValidator a -> Address

The address of the validator.

validatorScript :: TypedValidator a -> Validator

The unversioned validator script itself.

vValidatorScript :: TypedValidator a -> Versioned Validator

The validator script itself.

forwardingMintingPolicy :: TypedValidator a -> MintingPolicy

The unversioned minting policy that forwards all checks to the instance's validator

vForwardingMintingPolicy :: TypedValidator a -> Versioned MintingPolicy

The minting policy that forwards all checks to the instance's validator

forwardingMintingPolicyHash :: TypedValidator a -> MintingPolicyHash

Hash of the minting policy that forwards all checks to the instance's validator

generalise :: forall a. TypedValidator a -> TypedValidator Any

Generalise the typed validator to one that works with the Data type.