Plutus.Script.Utils.Scripts

Description

This module contains functions related to versioning scripts and BuiltinData, or more specifially, Datums and Redeemers. These functions do not depend on a particular version of Plutus.

Plutus language versioning

data Language

Non-Native Script language. This is an Enumerated type. This is expected to be an open type. We will add new Constuctors to this type as additional Non-Native scripting language as are added. We use an enumerated type for two reasons. 1) We can write total functions by case analysis over the constructors 2) We will use DataKinds to make some datatypes indexed by Language For now, the only Non-Native Scriting language is Plutus We might add new languages in the futures.

Note that the the serialization of Language depends on the ordering.

Constructors

PlutusV1

PlutusV2

data Versioned script

A script of some kind with its Plutus language version

Constructors

Versioned

Fields

unversioned :: script

version :: Language

Script hashing

scriptHash :: Versioned Script -> ScriptHash

Hash a Versioned Script

validatorHash :: Versioned Validator -> ValidatorHash

Hash a Versioned Validator script.

mintingPolicyHash :: Versioned MintingPolicy -> MintingPolicyHash

Hash a Versioned MintingPolicy script.

stakeValidatorHash :: Versioned StakeValidator -> StakeValidatorHash

Hash a Versioned StakeValidator script.

Script utilities

scriptCurrencySymbol :: Versioned MintingPolicy -> CurrencySymbol

The CurrencySymbol of a MintingPolicy.

Script data hashes

data Datum

Datum is a wrapper around Data values which are used as data in transaction outputs.

data DatumHash

Script runtime representation of a Digest SHA256.

data Redeemer

Redeemer is a wrapper around Data values that are used as redeemers in transaction inputs.

data RedeemerHash

Script runtime representation of a Digest SHA256.

datumHash :: Datum -> DatumHash

Hash a 'PV1.Datum builtin data.

redeemerHash :: Redeemer -> RedeemerHash

Hash a Redeemer builtin data.

dataHash :: BuiltinData -> BuiltinByteString

Hash a BuiltinData

Orphan instances

Pretty Language

Instance details

Methods

pretty :: Language -> Doc ann

prettyList :: [Language] -> Doc ann

Serialise Language

Instance details

Methods

encode :: Language -> Encoding

decode :: Decoder s Language

encodeList :: [Language] -> Encoding

decodeList :: Decoder s [Language]