Plutus.Script.Utils.V1.Contexts

Documentation

spendsOutput :: TxInfo -> TxId -> Integer -> Bool

Check if the pending transaction spends a specific transaction output (identified by the hash of a transaction and an index into that transactions' outputs)

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

The CurrencySymbol of the current validator script.

valueProduced :: TxInfo -> Value

Get the total value of outputs produced by this transaction.

valueSpent :: TxInfo -> Value

Get the total value of inputs spent by this transaction.

pubKeyOutputsAt :: PubKeyHash -> TxInfo -> [Value]

Get the values paid to a public key address by a pending transaction.

valueLockedBy :: TxInfo -> ValidatorHash -> Value

Get the total value locked by the given validator in this transaction.

scriptOutputsAt :: ValidatorHash -> TxInfo -> [(DatumHash, Value)]

Get the list of TxOut outputs of the pending transaction at a given script address.

fromSymbol :: CurrencySymbol -> ValidatorHash

Convert a CurrencySymbol to a ValidatorHash

ownHash :: ScriptContext -> ValidatorHash

Get the hash of the validator script that is currently being validated.

ownHashes :: ScriptContext -> (ValidatorHash, DatumHash)

Get the validator and datum hashes of the output that is curently being validated

pubKeyOutput :: TxOut -> Maybe PubKeyHash

Get the public key hash that locks the transaction output, if any.

txSignedBy :: TxInfo -> PubKeyHash -> Bool

Check if a transaction was signed by the given public key.

getContinuingOutputs :: ScriptContext -> [TxOut]

findContinuingOutputs :: ScriptContext -> [Integer]

Finds all the outputs that pay to the same script address that we are currently spending from, if any.

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

findDatumHash :: Datum -> TxInfo -> Maybe DatumHash

Find the hash of a datum, if it is part of the pending transaction's hashes

findDatum :: DatumHash -> TxInfo -> Maybe Datum

Find the data corresponding to a data hash, if there is one

findOwnInput :: ScriptContext -> Maybe TxInInfo

Find the input currently being validated.

data TxInInfo

An input of a pending transaction.

Constructors

TxInInfo

Fields

txInInfoOutRef :: TxOutRef

txInInfoResolved :: TxOut

data ScriptPurpose

Purpose of the script that is currently running

Constructors

Minting CurrencySymbol

Spending TxOutRef

Rewarding StakingCredential

Certifying DCert

data TxInfo

A pending transaction. This is the view as seen by validator scripts, so some details are stripped out.

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

Transaction inputs

txInfoOutputs :: [TxOut]

Transaction outputs

txInfoFee :: Value

The fee paid by this transaction.

txInfoMint :: Value

The Value minted by this transaction.

txInfoDCert :: [DCert]

Digests of certificates included in this transaction

txInfoWdrl :: [(StakingCredential, Integer)]

Withdrawals

txInfoValidRange :: POSIXTimeRange

The valid range for the transaction.

txInfoSignatories :: [PubKeyHash]

Signatures provided with the transaction, attested that they all signed the tx

txInfoData :: [(DatumHash, Datum)]

txInfoId :: TxId

Hash of the pending transaction (excluding witnesses)

data ScriptContext

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

scriptContextPurpose :: ScriptPurpose

newtype TxId

A transaction ID, using a SHA256 hash as the transaction id.

Constructors

TxId

Fields

getTxId :: BuiltinByteString

data TxOutRef

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

Constructors

TxOutRef

Fields

txOutRefId :: TxId

txOutRefIdx :: Integer

Index into the referenced transaction's outputs

data TxOut

A transaction output, consisting of a target address, a value, and optionally a datum hash.

Constructors

TxOut

Fields

txOutAddress :: Address

txOutValue :: Value

txOutDatumHash :: Maybe DatumHash

outputsAt :: Address -> TxInfo -> [Value]

Get the values paid to a public key address by a pending transaction.

valuePaidTo :: TxInfo -> Address -> Value

Get the total value paid to a public key address by a pending transaction.